added some changes

+++ RestaurantMenu/MenuItem.cs
@@ -6,12 +6,12 @@ namespace RestaurantMenu
 {
     class MenuItem
     {
-        public float Price { get; set; }
+        public double Price { get; set; }
         public string Description { get; set; }
         public string Category { get; set; }
         public bool IsNew { get; set; }
 
-        public MenuItem(float price, string description, string category, bool isNew)
+        public MenuItem(double price, string description, string category, bool isNew)
         {
             Price = price;
             Description = description;
